runfix: always calculate the next fire date
