+++ src/script/E2EIdentity/Enrollment.store.ts
@@ -21,20 +21,16 @@ import {QualifiedId} from '@wireapp/api-client/lib/user';
 import {constructFullyQualifiedClientId} from '@wireapp/core/lib/util/fullyQualifiedClientIdUtils';
 
 const e2eActivatedAtKey = 'e2eActivatedAt';
-const e2eTimer = 'e2eTimer';
 
 interface EnrollmentStore {
   store: {
     e2eiActivatedAt: (time: number) => void;
-    timer: (time: number) => void;
   };
   get: {
     e2eiActivatedAt: () => number;
-    timer: () => number;
   };
   clear: {
     deviceCreatedAt: () => void;
-    timer: () => void;
   };
 }
 
@@ -45,15 +41,12 @@ export const getEnrollmentStore = ({id: userId, domain}: QualifiedId, clientId: 
   return {
     store: {
       e2eiActivatedAt: (time: number) => localStorage.setItem(constructKey(e2eActivatedAtKey), String(time)),
-      timer: (time: number) => localStorage.setItem(constructKey(e2eTimer), String(time)),
     },
     get: {
       e2eiActivatedAt: () => Number(localStorage.getItem(constructKey(e2eActivatedAtKey))),
-      timer: () => Number(localStorage.getItem(constructKey(e2eTimer))),
     },
     clear: {
       deviceCreatedAt: () => localStorage.removeItem(constructKey(e2eActivatedAtKey)),
-      timer: () => localStorage.removeItem(constructKey(e2eTimer)),
     },
   };
 };
