+++ src/script/hooks/useIsSelfWithinGracePeriod.ts
@@ -1,0 +1,40 @@
+/*
+ * Wire
+ * Copyright (C) 2024 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+import {useCallback, useEffect, useState} from 'react';
+
+import {TIME_IN_MILLIS} from 'Util/TimeUtil';
+
+import {E2EIHandler} from '../E2EIdentity';
+
+export const useIsSelfWithinGracePeriod = () => {
+  const [isGracePeriod, setIsGracePeriod] = useState<boolean>(false);
+
+  const refreshGracePeriod = useCallback(async () => {
+    return E2EIHandler.getInstance().isWithinGracePeriod().then(setIsGracePeriod);
+  }, []);
+
+  useEffect(() => {
+    void refreshGracePeriod();
+
+    setTimeout(refreshGracePeriod, TIME_IN_MILLIS.SECOND);
+  }, [refreshGracePeriod]);
+
+  return isGracePeriod;
+};
